New Makefile system
There is a new directory called folder. It holds templates for
Makefiles. There are two type of Makefiles: libraries (as engine)
and modules (inside libraries). The purpose  of modules is to generate
object files (.o) and the purpose of libraries is to generate .a files

The library is stored into upper lib directory and automatically linked
by the general Makefile.

So, as of now, to initiate a new library or module link (ln -s) the
utils/make template and write a .options file to specify TARGET or
VERSION.